Lecture5/mybatis/src/main/java/pe/example/mybatis/dao/BoardDao.java
package pe.example.mybatis.dao;

import org.apache.ibatis.session.SqlSession;
import org.apache.ibatis.session.SqlSessionFactory;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Repository;
import pe.example.mybatis.dto.BoardDto;
import pe.example.mybatis.mapper.BoardMapper;

/**
 * @author D0Loop
 * @since 2022-02-26 [2022.2월.26]
 */

@Repository
public class BoardDao {
    private final SqlSessionFactory sessionFactory;

    public BoardDao(@Autowired SqlSessionFactory sessionFactory) {
        this.sessionFactory = sessionFactory;
    }

    public int createBoard(BoardDto dto) {
        try (SqlSession session = sessionFactory.openSession()){
            BoardMapper mapper = session.getMapper(BoardMapper.class);
            return mapper.createBoard(dto);
        }
    }

    public BoardDto readBoard(int id) {
        try (SqlSession session = sessionFactory.openSession()){
            BoardMapper mapper = session.getMapper(BoardMapper.class);
            return mapper.readBoard(id);
        }
    }
}

